aws_cdk_serverless_container_fargate/requirements.txt
-e .
aws.cdk.aws_ec2
aws.cdk.aws_ecs 
aws.cdk.aws_ecs_patterns